3d: Add logic to highlight a vector entity
This introduces `QgsAbstractVectorLayer3DHighlightFactory` which adds the
logic to crate an entity based on a feature and its 3d renderer. The
main idea is to replace the existing symbol associated with the
renderer by a new symbol where the colors are changed to highlight
the entity.

Each renderer needs to implement the
`QgsAbstractVectorLayer3DHighlightFactory::create` method to add the
implement the entity creation logic.

In pratice, the factory is created by calling the
`QgsAbstractVectorLayer3DRenderer::createHighlightFactory` method. And
then calling `create` on the factory to highlight a vector feature.