qgsrubberband3d: Do not set the parent on a component constructor
This ensures to avoid crashes if the entity is quickly deleted or
updated.

Indeed, Qt3D needs to keep both a frontend a backend in order to
optimize the scene graph interactions. This can result in some
entities being created and deleted in the frontend before the backend
is able to handle those entities.

Besides the parent logic is a mix of the Qt parent logic of any
QObject and the parent/child logic of the scene graph.

A simple rule to prevent any crashes is the following:
 - use a parent for entities in the constructor
 - do not use parents for components. An entity takes reponsability of
 of a component once added to the entity with the `addComponent` call.

This rule is applied to the |mLineEntity` entity. This should help
resolve segfaults that sometimes happen when using a rubberband.